requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
docstring-parser==0.3
entrypoints==0.3
Flask==1.1.1
funcsigs==1.0.2
graphviz==0.12
idna==2.8
imageio==2.5.0
importlib-metadata==0.23
instaboost==0.2
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.1.1
jupyter-client==5.3.4
jupyter-core==4.6.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nb-conda-kernels==2.2.2
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.5
notebook==6.0.1
numpy==1.17.0
opencv-mat==0.1
opencv-python==4.1.0.25
paddlehub==1.1.1
paddlepaddle-gpu
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prefetch-generator==1.0.1
prettytable==0.7.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.1
ptyprocess==0.6.0
pycocotools==2.0.0
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
pyzmq==18.1.0
rarfile==3.0
recordio==0.1.7
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.34.0
traitlets==4.3.3
ujson==1.35
urllib3==1.25.3
visualdl==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
wget==3.2
zipp==0.6.0
